Add files via upload
it turns out that the reason for the daily_training errors looks to be due to how the new changelog.json doesn't have an "entries" key in the dictionary. This was probably taken out by cline when retraing the llm from scratch, but since we are not building an llm from scratch over and over again I think just adding this key to the dictionary list will be beneficial for now.